+++ mycodo/mycodo_flask/routes_calibration.py
@@ -688,13 +688,16 @@ def setup_ds_resolution():
             return redirect(url_for('routes_admin.admin_dependencies',
                                     device='CALIBRATE_DS_TYPE'))
 
-    # If DS18B20 inputs added, compile a list of detected inputs
+    # If DS inputs exist, compile a list of detected inputs
     ds_inputs = []
     try:
         if os.path.isdir(PATH_1WIRE):
             for each_name in os.listdir(PATH_1WIRE):
                 if 'bus' not in each_name:
-                    ds_inputs.append(each_name)
+                    input_dev = Input.query.filter(
+                        Input.location == each_name).first()
+                    if input_dev:
+                        ds_inputs.append((input_dev.device, each_name))
     except OSError:
         flash("Unable to detect 1-wire devices in '/sys/bus/w1/devices'. "
               "Make 1-wire support is enabled with 'sudo raspi-config'.",
@@ -705,27 +708,29 @@ def setup_ds_resolution():
             form_ds.device_id.data):
         try:
             from w1thermsensor import W1ThermSensor
-            input_dev = Input.query.filter(Input.unique_id == form_ds.device_id.data).first()
+            device_name = form_ds.device_id.data.split(',')[0]
+            device_id = form_ds.device_id.data.split(',')[1]
             input_type = None
-            if input_dev.device == 'DS18B20':
+            if device_name == 'DS18B20':
                 input_type = W1ThermSensor.THERM_SENSOR_DS18B20
-            if input_dev.device == 'DS18S20':
+            if device_name == 'DS18S20':
                 input_type = W1ThermSensor.THERM_SENSOR_DS18S20
-            if input_dev.device == 'DS1822':
+            if device_name == 'DS1822':
                 input_type = W1ThermSensor.THERM_SENSOR_DS1822
-            if input_dev.device == 'DS28EA00':
+            if device_name == 'DS28EA00':
                 input_type = W1ThermSensor.THERM_SENSOR_DS28EA00
-            if input_dev.device == 'DS1825':
+            if device_name == 'DS1825':
                 input_type = W1ThermSensor.THERM_SENSOR_DS1825
-            if input_dev.device == 'MAX31850K':
+            if device_name == 'MAX31850K':
                 input_type = W1ThermSensor.THERM_SENSOR_MAX31850K
             else:
-                flash("Unknown input type: {}".format(input_dev.device),
+                flash("Unknown input type: {}".format(device_name),
                       "error")
 
             if input_type:
-                sensor = W1ThermSensor(sensor_type=input_type, sensor_id=input_dev.location)
-                sensor.set_precision(
+                sensor = W1ThermSensor(
+                    sensor_type=input_type, sensor_id=device_id)
+                sensor.set_resolution(
                     form_ds.set_resolution.data, persist=True)
             flash("Successfully set sensor {id} resolution to "
                   "{bit}-bit".format(id=form_ds.device_id.data,